refactor: split goals.rs into goals/ sub-modules
Split the 513-line src/goals.rs into src/goals/ with:
- mod.rs (8 lines): re-exports all pub items
- types.rs (168 lines): GoalStatus, GoalUpdate, GoalRecord, goal_slug
- store.rs (258 lines): GoalStore trait, InMemoryGoalStore, FileBackedGoalStore
- seed.rs (121 lines): seed_default_goals + tests

All pub items re-exported from mod.rs so crate::goals::X paths unchanged.
Every file ≤400 lines. Follows src/identity/ pattern.
